Add files via upload
The purpose of this document is simple.  To explain to you the creation of a TCP protocol that creates a socket communication with headers between a server and multiple clients that want to send to the server a request at the same time. You will find two server header files . The one of them uses select . All the files have code with the programming language PYTHON .